remove "dummy_urls" code from admin.py
The "dummy_urls" code seems to cause some kind of race condition where
a link in the admin to a child model will sometimes link to the parent model
and sometimes link to the child. This behaviour seems
random - every time you reload the dev server one of the two behaviours
will occur. Simply deleting this code seems to have solved this problem.
More tests are required as this may have broken other functionality.